Title: 최솟값 3개, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 250127/최솟값 3개/top-3-smallest-number.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 priority_queue<int> pq;
 int n;
-int ans = -1;
+long long ans = -1;
 void getAns();
 int main() {
